refactor(api-protocol): make types loadable without triggering schema creation
