+++ libs/api/protocol/src/lib/infra/schema/communication/communication-message.schema.ts
@@ -1,10 +1,8 @@
 import { AutoMap } from '@automapper/classes';
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 
-import {
-	ProtocolEntryType,
-	ProtocolMessageEntryBaseDocument,
-} from '../protocol-entry-base.schema';
+import { ProtocolMessageEntryBaseDocument } from '../protocol-entry-base.schema';
+import { ProtocolEntryType } from '../protocol-entry-type';
 
 @Schema({ _id: false })
 export class CommunicationMessagePayloadDocument {
