+++ libs/api/protocol/src/lib/infra/schema/protocol-entry-type.ts
@@ -1,0 +1,10 @@
+export enum ProtocolEntryType {
+	COMMUNICATION_MESSAGE_ENTRY = 'COMMUNICATION_MESSAGE_ENTRY',
+	RESCUE_STATION_SIGN_ON_ENTRY = 'RESCUE_STATION_SIGN_ON_ENTRY',
+	RESCUE_STATION_UPDATE_ENTRY = 'RESCUE_STATION_UPDATE_ENTRY',
+	RESCUE_STATION_SIGN_OFF_ENTRY = 'RESCUE_STATION_SIGN_OFF_ENTRY',
+	// UNIT_STATUS_ENTRY = 'UNIT_STATUS_ENTRY',
+	// OPERATION_STARTED_ENTRY = 'OPERATION_STARTED_ENTRY',
+	// OPERATION_ENDED_ENTRY = 'OPERATION_ENDED_ENTRY',
+	// OPERATION_UNITS_UPDATED_ENTRY = 'OPERATION_UNITS_UPDATED_ENTRY',
+}
