+++ libs/api/protocol/src/lib/infra/schema/unit-partial.schema.ts
@@ -1,5 +1,6 @@
 import { AutoMap } from '@automapper/classes';
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
+import { Types } from 'mongoose';
 
 export enum UnitType {
 	REGISTERED_UNIT = 'REGISTERED_UNIT',
@@ -17,7 +18,7 @@ export class RegisteredUnitDocument extends UnitDocument {
 
 	@Prop()
 	@AutoMap()
-	unitId: string;
+	unitId: Types.ObjectId;
 }
 
 @Schema({ _id: false })
@@ -28,6 +29,8 @@ export class UnknownUnitDocument extends UnitDocument {
 	@AutoMap()
 	name: string;
 }
+
+export type UnitDocuments = RegisteredUnitDocument | UnknownUnitDocument;
 
 export const RegisteredUnitSchema = SchemaFactory.createForClass(
 	RegisteredUnitDocument,
