+++ libs/api/protocol/src/lib/mapper-profile/unit-partial.mapper-profile.ts
@@ -2,6 +2,7 @@ import type { Mapper } from '@automapper/core';
 import { createMap, forMember, mapFrom } from '@automapper/core';
 import { AutomapperProfile, getMapperToken } from '@automapper/nestjs';
 import { Inject, Injectable } from '@nestjs/common';
+import { Types } from 'mongoose';
 
 import {
 	RegisteredUnit,
@@ -65,7 +66,7 @@ export class UnitPartialProfile extends AutomapperProfile {
 				forMember(
 					(d) => d.unit,
 					mapFrom((s) => ({
-						id: s.unitId,
+						id: s.unitId.toString(),
 					})),
 				),
 			);
@@ -78,7 +79,7 @@ export class UnitPartialProfile extends AutomapperProfile {
 				RegisteredUnitDocument,
 				forMember(
 					(d) => d.unitId,
-					mapFrom((s) => s.unit.id),
+					mapFrom((s) => new Types.ObjectId(s.unit.id)),
 				),
 			);
 			createMap(mapper, UnknownUnit, UnknownUnitDocument);
