+++ libs/api/protocol/src/lib/protocol.module.ts
@@ -18,8 +18,8 @@ import { CommunicationMessageSchema } from './infra/schema/communication/communi
 import {
 	ProtocolEntryBaseDocument,
 	ProtocolEntryBaseSchema,
-	ProtocolEntryType,
 } from './infra/schema/protocol-entry-base.schema';
+import { ProtocolEntryType } from './infra/schema/protocol-entry-type';
 import { RescueStationSignOffMessageSchema } from './infra/schema/rescue-station/rescue-station-sign-off-message.schema';
 import { RescueStationSignOnMessageSchema } from './infra/schema/rescue-station/rescue-station-sign-on-message.schema';
 import { RescueStationUpdateMessageSchema } from './infra/schema/rescue-station/rescue-station-updated-message.schema';
